Separate Vector3 and Point3
These don't behave identically (though there's lots of overlap), so they
shouldn't be implemented with the same type.

I've added Point::add and Point::mul<f64> at PBR's behest; the latter is
currently used in Perlin noise, and the former is unused. There's more
to PBR's Point3 interface (e.g. linear interpolation) but I've left it
for now.

+++ src/geometry/vector/mod.rs
@@ -33,10 +33,10 @@ impl From<[f64; 3]> for Vector3 {
     }
 }
 
-impl Index<u64> for Vector3 {
+impl Index<usize> for Vector3 {
     type Output = f64;
 
-    fn index(&self, index: u64) -> &Self::Output {
+    fn index(&self, index: usize) -> &Self::Output {
         match index {
             0 => &self.x,
             1 => &self.y,
